Read multiline string from file instead of buiding with .join().
Summary:
Leverage our `ProjectWorkspace.getFileContents()` utility to read "multiline,
expected data with platform-specific differences" from a file rather than
building it programmatically with `Joiner` because the former is much easier to
read and maintain.

I'm going to be updating this test soon, so I wanted to clean this up first.

Test Plan: CI (pretty sure this should work on Windows)

fbshipit-source-id: 864ea66