Disable OSHI warnings going to LogView
Summary: We use OSHI for process tracking, and sometimes between the time the process was listed and queried for its stats, the process finishes which causes no such file exception for `/proc/pid/<finished_process>`. This doesn't fail the build or anything but OSHI logs this as Log.Error which causes it to go to LogView. We don't care about this so let's turn off all OSHI logging.

Test Plan:
Inserted snippet of code and verified that after the change it no longer gets logged.

    if (cnt++ < 5) {
      org.slf4j.Logger LOG = LoggerFactory.getLogger(FileUtil.class);
      LOG.error("plamenko test error {} {}", "file", new Exception("test exception"));
    }

Reviewed By: aiked

fbshipit-source-id: ad5a2fb