audit classpath: support option to use versioning with action graphs
Summary:
This uses the "build"-specific buck config entry which enables
versioning to gate whether versioning is used for audit classpath
commands which build the action graph.  This works well, as if
versioning is enabled for building it means `versioned_alias` rules
are present in the target graph, which cannot be transformed to
actions without versioning.

This also allows keeping versioning disabled for audit commands
which don't require an action graph, which is useful for speed
and debugging.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: e46e8bc