Rename TracesHelper to BuildTraces and move from httpserver to util.trace.
Summary:
Move `TracesHelper` into a common location so that it can be used by
`ProjectWorkspace` for testing (which will happen in a follow-up revision).
This revision was brought to you by IntelliJ and `buck autodeps`.

A number of formerly package-private APIs had to be made public so
that things would still build. Also, `TracesHandlerDelegate.TRACE_ID_PATTERN_TEXT`
was moved to `TracesHelper` to avoid a circular dependency.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: 2084c3e