Introduce ChromeTraceParser.
Summary:
For our integration tests, it would be nice to be able to extract ad-hoc
information about the invocation of Buck with minimal overhead. Fortunately, the
combination of the `BuckEventBus` and the `ChromeTraceBuildListener` already
make it possible to subscribe to any event posted to the event bus and record it
as an event in the `build.trace`. (One-off events should be tagged with phase
`"i"` to indicate an "instant" event.)

However, to extract this information, we need some sort of parser for a
`build.trace`, so to that end, this revision introduces `ChromeTraceParser`.
Because the `build.trace` may be a very large file, we do not want to load the
entire thing into memory only to extract a handful of events. This revision
generalizes some logic in `BuildTraces` to make it possible to provide an
arbitrary matcher for the events that you want to extract.

Although this code is currently focused on events that only need to be matched
once in the trace, this utility could be generalized to support more complex
conditions. I would rather wait to expand this logic once there is a use case to
motivate the a more complex API.

Test Plan: Created `ChromeTraceParserIntegrationTest`.

Reviewed By: illicitonion

fbshipit-source-id: 7ca9bef