Don't put source files in classpath for CompilerTreeApiTest
Summary:
`CompilerTreeApiTest.withClasspath` lets a test provide classes that
should be included on the classpath of the main compilation.

However, it was putting the `.class` files in the same directory as
their corresponding `.java` files, and using that directory as the
classpath for the main compilation. This was (in some cases, due
to something nondeterministic in `javac` that I didn't track down
fully) triggering a "helpful" behavior in `javac` whereby it
automatically compiles `.java` files from the classpath if it looks
like you're trying to reference them.

The fix is to make the test separate the source and class files.

Test Plan: Observe no more intermittent failures in CI for tests that use this functionality

Reviewed By: dreiss

fbshipit-source-id: b26f7e8