Validate type and constant references
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

When generating ABIs from source, we'll be running `javac`
with just the boot classpath and a few critical dependency modules.
Any simple names that can't be resolved in that environment we will
assume refer to classes in the same package as the current compilation
unit. Any constants that are defined in modules that aren't present will
appear not to be constant.

This commit adds a validator that detects type and constant references
which would not resolve properly during ABI generation.

Test Plan: New unit tests

Reviewed By: dreiss

fbshipit-source-id: 6a92f6a