Move unknown type handling to TypeResolver
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

Because we're resolving types without having access to all dependency
modules, if we see a type reference in the code that we can't resolve, we'll
we'll create a dummy type for it. This allows the ABI generator that will
eventually use this code to at least see the name of the type, which should
be pretty much all it needs.

However, contrary to the TODO in the code, `getTypeElement` is not the
place to create such dummy types. Annotation processors call `getTypeElement`,
and they're built to expect `null` returned when types don't exist. If an annotation
processor is calling `getTypeElement`, it's probably because it needs to inspect
the resulting type, so we want to detect that the type is missing and add it as
a dependency at ABI generation time. Always returning a dummy type will push
the failure point a little further on and make troubleshooting more difficult.

This commit moves the TODO to `TypeResolver`, which is only used to resolve
references that appear in the code. This required a small amount of restructuring.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 4c60b72