Generalize StubJar [1/4]: Factor out stub creation
Summary:
`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s. To support that, it needs to be less coupled to the assumption
that it's looking at bytecode.

Step 1: Factor out the code that invokes the ASM `ClassVisitor` to write the
stub. In the future we'll have another driver for it that starts from a
`TypeElement` rather than an `InputStream` over a `.class` file.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: ba0922d