Generalize StubJar [2/4]: Factor out unstubbed file access
Summary:
`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s. To support that, it needs to be less coupled to the assumption
that it's looking at bytecode.

Step 2: Delete `Walker`, replace with a pull model in `LibraryReader`. In the
future we'll have a version that works with a list of `TypeElement`s instead
of a directory listing.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 4cab9ef