Generalize StubJar [3/4]: Factor out jar writing
Summary:
`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s. To support that, it needs to be less coupled to the assumption
that it's looking at bytecode.

Step 3: Factor out the actual stub jar writing. In the future, we'll sometimes
use the `JavaFileManager` within `javac` to write the files instead of
`HashingDeterministicJarWriter`.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 3e1dcdf