Generalize StubJar [4/4]: Allow different types of unstubbed classes
Summary:
`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s. To support that, it needs to be less coupled to the assumption
that it's looking at bytecode.

Step 4: Enable `StubJar` to make use of `LibraryReader`s that return something other
than an `InputStream` when working with classes. We'll use this functionality shortly
to provide a `TypeElementsLibraryReader` that returns `TypeElement`s for classes.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 34285f1