Make migration mode issue warnings
Summary:
It was issuing mandatory warnings, which are fatal. That makes
migration en masse a bit of a pain. :-)

Test Plan:
Enable `abi_generation_mode=migrate_to_source` under `java` in `.buckconfig`
Build a project with migration issues
Observe those issues appear in the log as warnings but do not terminate the build

Reviewed By: dreiss

fbshipit-source-id: f42a5e8