See how many ActionGraph hits we'd get if we implemented TargetNode equality.
Summary:
This adds the code to track the cases where we currently get an
ActionGraphCache miss but where TargetGraph HashCodes are the same
(which indicates that the JSON the nodes were created from is the same).
This should allow us to get a decent idea of what the win of implementing
TargetNode equality would be.

Test Plan:
added new test
manual: run build, run build again, see ActionGraphCache hit, add whitespace to BUCK file, observe new event in log

Reviewed By: beefon

fbshipit-source-id: b45d2d9