Handle lack of C++ runtime libs in "system" runtime
Summary:
The "system" C++ runtime for Android is a bit different from the others
(gnustl and stlport).  It's runtime shared library is not part of the
toolchain, distributed with the app.  It's part of the platform,
distributed with the device.  As a result, we were failing to find the
runtime libs directory whenever attempting to build against this
runtime.

This diff updates the NDK platforms to treat the runtime libs directory
as optional and not set it for the system runtime.  The one point of use
was already guarded by an appropriate check.

Note that this does *not* update the linker to add the proper flags to
link against the libstdc++.so from the Android platform, but it is
enough to build a pure C cxx_library with no gnustl entanglements.

Test Plan: Ran exopackage test.

fbshipit-source-id: 518996b