Update Exopackage to recognize BUCK.fixture
Summary:
A while back, all of the BUCK files in test projects were renamed to
"BUCK.fixture" to avoid confusion.  The normal test runner automatically
renames them, but the exopackage test works a bit differently, so this
diff updates it to do the rename.

Test Plan: Ran it (with some other changes).

Reviewed By: illicitonion

fbshipit-source-id: cab5d9c