Make exopackage key formatting uniform between Java and native
Summary:
We have two versions of the Buck Android Agent (which is used for
exopackage installation): one in Java and one in C, because different
Android platforms have issues with each.  They were not formatting their
keys the same way.  Update the Java version to use the same format as
the C version (captial hex, with a newline terminator).

Test Plan: Ran the exopackage test (with some other changes).

fbshipit-source-id: 49700c1