Change the extension point for CompilerTreeApiTest
Summary:
One of the overloads of `CompilerTreeApiTest.compile` used to be the
thing to override if you wanted to do something no matter how the
compiler was invoked. After `initCompiler` was added, one if its
overloads became a better place, but I neglected to make it so.

Now I need it in a new test, so here we are.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 2932ac1