Initial set of failing tests for stub jar generation from source
Summary:
`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

This commit adds parameters to the existing `StubJar` unit tests, allowing them
to compare the class-generated stub with the source-generated one (with and
without the classpath) for each of the test cases. This is not a complete set
of tests for generating ABIs from source (especially for the missing classpath
case), but it's a good first target.

Test Plan:
Currently all of the tests are disabled when run against source-generated stubs.
They still run against class-generated stubs. They'll get enabled against source-generated
stubs as implementation proceeds.

Reviewed By: dreiss

fbshipit-source-id: c1ffb96