Stub jars from source: skip private members
Summary:
`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

This commit makes the stub generation not invoke the `ClassVisitor` for
private members. The `ClassVisitor` implementation was already ignore private members,
but this is more efficient (and avoids an NPE in the method case).

Test Plan:
Enabled the test for skipping private methods (the private fields one was already
passing thanks to `ClassMirror`).

Reviewed By: dreiss

fbshipit-source-id: 119cdfc