empty diff does not cause VCS failure
Summary:
It is possible that when the user invokes a buck rage there is no actual diff between the tip and base revision.
That would make rage fail the vcs gathering. This way, for this specific "error" it will not fail and proceed to
produce the rest of the data.

The other minor change of the changedFiles is a lint warning that I did not want to fix on another commit.

Test Plan: -CI

Reviewed By: beefon, illicitonion

fbshipit-source-id: f160621