Compare project filesystems by root in PrefixMapDebugPathSanitizer
Summary:
It looks like there's a case where the equality check would fail due to
action graph caching.  I wasn't able to repro it, but this semantic
check should be safe in any case (if slightly slower), because the check
is really to make sure that the roots are the same.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 6bd48b1