Fix JUnit test runner to address issues around classloading.
Summary:
Fixes a problem that occurs when the JUnit java_test binary includes a class that fails to classload in the classloader used by the JUnitRunner main method.  Before running tests from a given class, the JUnit test runner first *loaded* the class and inspected it to see if it looks like a test (e.g., if it has any test methods, if
it's annotated with RunWith, etc.)  If the class looked like a test, then we attempted to run it as a test.  If the class failed to classload, the java_test was aborted, and no further tests were run.

The problem we witnessed was a class failing to load because it could not link at runtime.  This was because a method signature recorded at compile-time was not available at runtime, which was generating an uncaught LinkageError that caused our JUnitRunner to immediately exit without attempting any more tests.

This fix catches that exception and reports it as a test failure if the class looks like the user intended it to be a test class (meaning that it has "Test" somewhere in the base part of its class name).  Otherwise, the failure is ignored.

Test Plan: Added unit test.

Reviewed By: asp2insp

fbshipit-source-id: 66ce760