Refactor dep query code into it's own class
Summary:
This re-organizes the code for dep query into it's own class with
it's own type coercer.  This will make it easier in future changes
to have it automatically do the right thing when used in a constructor
arg (e.g. add itself to parse-time deps, rewrite it's targets correctly
for versioning).

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 1ddd9e9