If Watchman returns a fresh instance, update the clock
Summary: We return before updating the clock when we get an overflow event due to a new watchman instance. We should update the clock accordingly.

Test Plan:
Need to figure out what this actually means and then test it.

Worst case:

1. We see an overflow event
2. We update the clock to NULL clock
3. Query way back in time and get an overflow event
4. Goto 2

I don't think that's what should happen, but I want to test it

Reviewed By: yiding

fbshipit-source-id: eaad252