versions: add interface for adding custom target translators
Summary:
Previously, implementations for translating build targets in various
types were put directly into `TargetNodeTranslator` or in the type
itself (for types which, for whatever reason, couldn't use the
`TargetNodeTranslatable` interface).  This diff instead adds a new
interface which allows the implementations to be moved to their own
classe and registered in the `TargetNodeTranslator` externally.

Test Plan: added test

Reviewed By: illicitonion

fbshipit-source-id: d266744