Allow bridging header import headers under current module
Summary:
Right now when buck building swift with bridging header, it is not able to import header files when it reference from current library.

For example, in build rule with this setup

```
apple_library(
    name = 'Greeter',
    bridging_header = 'bridging-header.h',
    headers = glob([
        '*.h',
    ]),
    srcs = glob([
        '*.m',
        '*.swift',
    ]),
    frameworks = [
        '$SDKROOT/System/Library/Frameworks/Foundation.framework'
    ],
)
```

We can not do `#import <Greeter/Greeter.h>` in bridging header.

After investigation, I figure its due to buck didn't include the header map generated for current library in SwiftCompile.

This PR is to improve buck so it can support above use case.

**Testing**
And a new test case and verify it with `buck test --all`
Closes https://github.com/facebook/buck/pull/1164

Reviewed By: ryu2

fbshipit-source-id: 3a2b63c