Cache ParamInfo in TargetNodeFactory
Summary:
ParamInfo.<init> does reflection, which is slow.  Cache the ParamInfo
objects in the TargetNodeFactory to avoid repeating it for every
Description type.  This should be safe because the ParamInfo only
depends on the UnpopulatedConstructorArg class (which is the cache key)
and the typeCoercerFactory, which is also scoped to the
TargetNodeFactory.

Test Plan: CI.  Action graph construction is several percent faster.

Reviewed By: marcinkosiba

fbshipit-source-id: 333ef90