Move c.f.b.android.support to its own directory
Summary:
This isn't really part of buck proper and we've been forced to ignore
it in our intellij project settings. This moves it to its own directory
and configures intellij such that we actually get indexing and all
kinds of real features when editing code here.

Test Plan: CI, open in intellij and edit one of these files

Reviewed By: dreiss

fbshipit-source-id: c4c2973