Add forEachFwdDep to QueryEnvironment
Summary:
Previously, DepsFunction was using getFwdDeps to iterate over all deps.
However, this requires constructing an intermediate set, which is a
little slow.  forEachFwdDep avoids this by passing the visitor all the
way down to the action graph nodes.  The downside is that the visitor is
invoked multiple times on some nodes, but it is so cheap in this case
that it's still a net win.

Test Plan: CI.  Saw action_graph be faster for fb4a.

Reviewed By: asp2insp

fbshipit-source-id: ad6a1aa