Add `js_file` rule
Summary:
Port the `js_file` rule from build defs to Java.

This rule is used to transpile a single JavaScript source file with a `worker_tool`. `#prod` is a special flavor that post-processes the result of the same target without the `prod` flavor. I.e. `#prod` always depends on the non-`prod` version of the same target.

`extraArgs` will be used later to control additional command line flags for the worker based on flavors configured in `.buckconfig`

`virtualPath` is specifically needed for `node_modules`. We install them into `buck-out`, but the packager needs their ?virtual location? to be in the source tree for proper dependency resolution.

Flavors are currently hardcoded, but will be configurable eventually, with the notable exception of `prod`.

Test Plan: new integration test

Reviewed By: marcinkosiba

fbshipit-source-id: 858a3af