Don't ignore field structure in dep-file RKF
Summary:
Instead of first collecting SourcePaths and then processing them afterwards, SourcePaths are processed immediately. That way the field name and structure is encoded in the rulekey rahter than being lost in the process of dumping all SourcePaths in a Set. This should make dep-file rule keys more correct, and should avoid some unwanted collisions that were possible before.

We used to process and cache RuleKeyAppendables separately, but this gets tricky now, so I dropped the appendable cache completely. The difficulty is that the same appendable may get processed differently when referenced from different rules. In particular, if appendable references a SourcePath, how this path gets handled depends on the rule itself. Before this was not a problem because we would just gather a list of SourcePaths, and would only process it at the end.
I also noticed that dep-file factory don't perform caching on per rule level because it was completely bypassing the logic of its parent class :(

So I performed a couple of benchmarks to see what is the perf effect of per-rule and per-appendable caching. All builds are done on a 15" MacBook Pro with SSD, with `rm -rf buck-out`, `rm -rf buck-cache`, `--deep` and disabled http cache. Nothing else (apart from standard MacOSX background processes) was being done on the laptop during the build (i.e. I was not surfing). Results are attached.

In short, the overall build time doesn't seem to be significantly affected by this change. Neither appendable, nor per-rule cache seems to help. In fact, per-rule cache makes things worse for some reason. (Possibly because of additional memory and cpu pressure of cache bookkeeping, whereas the cache is not really utilized).

Test Plan: performed benchmarks, ran unit & integration tests

Reviewed By: illicitonion, Coneko

fbshipit-source-id: a870279