Manually specify which classes should be stubbed in tests
Summary:
Previously we had some logic in `assertClassesStubbedCorrectly` that tried
to figure out which classes should be stubbed. That logic largely duplicated
the logic in the product code. The product code is about to get really
complicated as we try to include technically `private` classes that are
nonetheless part of the ABI (e.g. superclasses of public classes), so
this commit changes the tests to manually specify which classes should
be stubbed or not.

Test Plan: Updated unit tests

Reviewed By: dreiss

fbshipit-source-id: c736343