Fix validator crashing on unknown type
Summary:
The validator was crashing on unknown types due to them coming back
from the reference finder in non-fully-qualified form.

This commit fixes the issue deeply, by both enabling the validator to
deal with types in the default package, and changing the reference finder
to only report properly resolved types.

This involved reworking the `InterfaceTypeAndConstantReferenceFinderTest`
to be able to work with code that doesn't compile.

Test Plan:
New unit tests
Introduce a reference to a missing type in Buck code, verify no crash on `buck build`

Reviewed By: dreiss

fbshipit-source-id: 55f7f5d