Make AaptPackageResources calculate its own deps
Summary:
This moves the logic of calculating AaptPackageResources deps from the
graph enhancer into AaptPackageResources itself. It's a lot easier to
reason about the correctness of the deps when it is alongside the code
that actually uses things from deps.

Test Plan: CI, added test that aaptpackageresources rulekey reflects things it should

Reviewed By: illicitonion

fbshipit-source-id: 02fecd5