query: new optional argument to `deps()` to control traversal
Summary:
This diff adds a new optional argument to the `deps()` function which
is an expression which can be used to determinie whether a node is
traversed in the dep search.  To provide a way to reuse existing
functions expressions, this also adds a new `first_order_deps()` function
which can only be used from the context of the filter expression argument
to `deps()`, which resolves to the first-order deps of current node being
inspected during traversal.  This allowing composing expressive filter
expressions such as `buck query 'deps(//foo:bar, 10, kind(.*_library, first_order_deps()))`
which expand resolves to the library deps of `//foo:bar` (note that this
is different from the query `kind(.*_library, deps(//foo:bar, 10))`
which includes libraries found from crossing non-library deps).

Test Plan: added tests

Reviewed By: asp2insp

fbshipit-source-id: c797ad4