Override -proc:only for DummyRDotJava
Summary:
The behavior of -proc:only changes in javac 1.8 such that more symbols end up being completed/resolved.  The 'R' class is one example.  When buck creates the dummy 'R.java' in a library_module() that uses annotation_process_only=True, buck internally generates R.java but the compilation with -proc:only does not output the R.class to the resulting .jar file.  The actual library compilation then fails with javac 1.8 because it tries to resolve 'R' that does not exist.  In javac 1.7 with -proc:only, javac did not try to resolve 'R', which is why this worked before.

Traditionally we use exactly the same options to compile both R.java and the rest of the library.  This overrides -proc:only for the dummy R.java jar we produce so that the real library can compile under javac 1.8.

Test Plan: add unit test

Reviewed By: dreiss

fbshipit-source-id: b394d7d