A more local fix for the validator crashing on unknown types
Summary:
The previous attempt at fixing this issue was a deep fix, fixing both
the root cause and the symptom. The root cause part uncovered an issue
with our continuous testing and was reverted. I'm still working through
that issue, so this is the symptom-fixing part.

Test Plan: Introduce a reference to a missing type in Buck code, verify no crash on buck build

Reviewed By: dreiss

fbshipit-source-id: 9964060