Prevent annotation processor auto-discovery in CompilerTreeApiTest
Summary:
`CompilerTreeApiTest` was not specifying any annotation processors when
invoking the compiler. This resulted in the compiler doing auto-discovery,
which found the Immutables processor and used it. That resulted in extra
processing which both slows the tests down a bit and created noise
in tests that were not expecting it.

Test Plan: Existing unit tests derived from this class

Reviewed By: dreiss

fbshipit-source-id: c2892f2