Abort a build if the sparse profile import fails
Summary:
The log will contain the gory details, but the user needs to know that their
build may fail if hg sparse exits with an exception. Use both a severe console
log entry and an exception for extra visibility.

Move ConsoleEvent to event_interfaces to avoid a circular dependency.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 940af32