Test runner should return non-zero on crash
Summary:
The Java test runner was always exiting with a code of 0,
even if it crashed while running tests. This is OK when
the runner was being used directly by Buck, because Buck
would look at the test result files and notice that some
tests were not run. External test runners have no such
requirement, so some of them were not noticing that
some tests did not run.

Test Plan:
Run a test that leads to the runner crashing (e.g. create
a test class that fails to load because of referencing
a type that's not present on its runtime classpath). Observe
both Buck and external test runners properly fail.

Reviewed By: marcinkosiba

fbshipit-source-id: 9d6aeb8