Make CalculateAbi.FLAVOR private 1/3: Checking for ABI targets
Summary:
We're going to start generating different ABI jars to use depending on
whether a module is in the same package as its dependent.

There are a few places in the code where we assume only one ABI jar target
exists per library target, and those will need to go away. This series of
commits makes `CalculateAbi.FLAVOR` private to remove one category of such
assumptions.

There are three scenarios for using this constant. This commit deals
with the places where it is used to determine if a target is an ABI target by
calling a new `CalculateAbi.isAbiTarget` function in each of those places.

Test Plan: New unit test

Reviewed By: marcinkosiba

fbshipit-source-id: ca0e450