Make CalculateAbi.FLAVOR private 3/3: Finding the ABI target for a library target
Summary:
We're going to start generating different ABI jars to use depending on
whether a module is in the same package as its dependent.

There are a few places in the code where we assume only one ABI jar target
exists per library target, and those will need to go away. This series of
commits makes `CalculateAbi.FLAVOR` private to remove one category of such
assumptions.

There are three scenarios for using this constant. This commit deals
with the places where it is used to navigate from a library target to its
corresponding ABI target.

Test Plan: New unit test

Reviewed By: marcinkosiba

fbshipit-source-id: 5027efb