Centralize determination of ABI target name
Summary:
We're going to start generating different ABI jars to use depending on
whether a module is in the same package as its dependent.

There are a few places in the code where we assume only one ABI jar target
exists per library target, and those will need to go away.

Every rule that built a jar had its description computing the ABI
target name and passing it to the constructor. Java 8 default methods
let us do this on the `HasJavaAbi` interface instead. Doing so will
make it a lot easier to add a second ABI target.

Test Plan:
Moved unit tests with the code
CI for regressions

Reviewed By: marcinkosiba

fbshipit-source-id: 5ad1a44