Allow cxx rules to depend only on headers of another rule
Summary:
This enhances CxxLibrary Noop rule to allow a "header only" version of some
rule to be instantiated when the no-op rule is built with only a "#header" (and
no platform) flavor.

Motivation:
This allows library consumers to separate the dependence on another library's
exported headers for compiling source files from the library's compiled output
used for linking further up the dependency graph at the binary rule.

This allows for use cases such as having code compile against some library, but
at link time link-in a different library that provides the symbols.

```
For this dependency graph:

BinaryA --depends-on--> LibA#headers, LibAB#shared
BinaryB --depends-on--> LibB#headers, LibAB#shared

LibAB --depends-on--> LibA, LibB

LibA --depends-on--> LibZ
LibB --depends-on--> LibZ

Results in action graph that looks something like this:

LibAB.so --re-exports--> LibA.a, LibB.a, LibC.a

BinaryA.exe --links--> LibAB.so (but not LibA.a, LibC.a)
BinaryB.exe --links--> LibAB.so (but not LibB.a, LibC.a)
BinaryA.o --uses-headers--> LibA.h, LibC.h (but not LibB's)
BinaryB.o --uses-headers--> LibB.h, LibC.h (but not LibA's)
```

Test Plan: Added an integration test

Reviewed By: Coneko

fbshipit-source-id: 931f290