Factor out annotation processor properties
Summary:
This commit creates `JavacPluginProperties`, a class that holds the
properties of a `javax.annotation.processing.Processor` or
`com.sun.source.util.Plugin`, as opposed to the parameters with
which it is invoked for a particular build rule. (The latter are
still held by `AnnotationProcessingParams`.)

As we move forward toward a `java_plugin` rule instead of per-library
`annotation_processors` and `annotation_processor_deps` properties,
`JavacPluginProperties` will be computed once per `java_plugin` and
cached, saving us time during action graph creation.

Test Plan: CI; there are a variety of annotation processing tests in `DefaultJavaLibraryIntegrationTest`

Reviewed By: asp2insp

fbshipit-source-id: 4f79b91