Make ClassLoaderCache easier to work with
Summary:
Previously, users of `CacheLoaderCache` had to be careful not to
close the `ClassLoader`s they received from the cache. This resulted in
silliness like `ProcessorBundle.closeClassLoader`.

This commit has `ClassLoaderCache` return `ClassLoader`s that ignore calls to
`close`, leaving a private method that the cache calls to close them when it's
time.

Test Plan: New unit tests

Reviewed By: dreiss

fbshipit-source-id: fc3c12b