Jsr199Javac: Treat annotation processors separately when evaluating safety
Summary:
Previously we would use a single `ClassLoader` to load all annotation processors
for a given build rule. Thus, in order for that `ClassLoader` to be reused across
rules, all annotation processors had to be on the safe list. This made the safe list
very fragile.

This commit lets us use up to two `ClassLoader`s per build rule -- one for the safe
processors (which is owned by the global `ClassLoaderCache` and thus lives across rules)
and one for the unsafe ones (owned by a per-rule `ClassLoaderCache`, so it is only used
for one compilation).

In addition to making the existing safe list less fragile, this approach to obtaining
`ClassLoader`s is more directly aligned with the coming `java_plugin` rule, which will
allow each processor to have its own classpath.

Test Plan: Updated unit tests, various annotation processing tests in `DefaultJavaLibraryIntegrationTest`

Reviewed By: asp2insp

fbshipit-source-id: 715c6f4