java_annotation_processor
Summary:
This commit adds a new `java_annotation_processor` rule which holds metadata about
a particular annotation processor, and a `plugins` parameter on `java_library` for
specifying which compiler plugins (of which annotation processors are one kind) to
use in compiling the library.

Future commits will add a `javac_plugin` rule specifically for the new plugin
support in `javac` 8 (`com.sun.source.util.Plugin`). The `plugins` parameter will
accept either, on the grounds that people shouldn't have to care which kind of plugin
they're dealing with. It is unfortunate that the `Plugin` class is so generically
named. :-(

Test Plan:
Migrated all uses of the old `annotation_processors` parameter in Buck's own build
and test fixtures to `plugins`. Added a test for query.

Reviewed By: dreiss

fbshipit-source-id: 0ff060b