Test for proper stubbing of static member classes, enums, and interfaces
Summary:
(Reapplying this commit after the stack was reverted; this
commit was not related to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

This commit fills some test holes by testing the stubbing of static members,
enums, and interfaces, fixing a bug discovered in the process.

Test Plan: New unit test

Reviewed By: asp2insp

fbshipit-source-id: 1c6fcc4