Stub jars from source: Properly generate enclosing instance pointer for deeply nested classes
Summary:
(Reapplying this commit after the stack was reverted; this
commit was not related to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

This commit fixes the previously hardcoded naming for the enclosing instance pointer.

Test Plan: New unit test

Reviewed By: asp2insp

fbshipit-source-id: d0522d6