Stub jars from source: include throws clauses
Summary:
(Reapplying this commit after the stack was reverted; this
commit was not related to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

This commit enables proper output of `throws` clauses in source-based stub jars.

Test Plan: New unit test

Reviewed By: asp2insp

fbshipit-source-id: fef844a