Stub type annotations
Summary:
(Reapplying this commit after the stack was reverted; this
commit was not related to the failure.)

Type annotations are some new Java 8 hotness that our ABI-generation
code was ignoring.

This commit teaches the jar-based ABI generator to stub type annotations.
For now the source-based generator is left out, because `javax.model` doesn't
provide easy access to type annotations. As the source-based generator gets
used in heavy Java 8 environments, we'll fix that up.

Test Plan: New unit tests

Reviewed By: asp2insp

fbshipit-source-id: 4e5eaac