Add some tests for annotation values of type Class
Summary:
(Reapplying this commit after the stack was reverted; this
commit was not related to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

This commit closes some test holes. No bugs found, but good to have the coverage.

Test Plan: New unit tests

Reviewed By: asp2insp

fbshipit-source-id: 4638e1d