Stub jars from source: Correct visibility of inner classes
Summary:
(Reapplying this commit after the stack was reverted; this
commit was not related to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

It's kinda fun how inner classes got added to the format. Looks like the
class file format still only allows the `public` access flag at the file level
(or no flag for default visibility), and if you wanted the class to actually be
`protected` or `private` that gets encoded in the `InnerClasses` attribute.

Test Plan: New unit tests

Reviewed By: asp2insp

fbshipit-source-id: 012f671