Don't stub <clinit>
Summary:
(Reapplying this commit after the stack was reverted; this commit was not related
to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

There was a test failing because the source-generated stub didn't have a
`<clinit>` method for a class with a class initializer. However, such methods
don't need to be present in order to compile code that uses the class, so
the easy way out is to just not emit `<clinit>` in stubs at all.

Test Plan:
Added a new test to enforce the new behavior
Enabled a new source-based stub unit test
CI for regressions in other areas

Reviewed By: dreiss

fbshipit-source-id: b66576b