Add some tests for different annotation retentions
Summary:
(Reapplying this commit after the stack was reverted; this	commit was not related
to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

This commit closes some holes in the testing. No bugs found, but good to
have the coverage.

Test Plan: New unit tests

Reviewed By: dreiss

fbshipit-source-id: 0222c8b