Simplify how new stub jar tests are written
Summary:
(Reapplying this commit after the stack was reverted; this	commit was not related
to the failure.)

`StubJar` is going to be gaining the ability to generate stub jars based on
`TypeElement`s.

We're going to need to write a lot more tests; the existing ones have holes,
and also there's no coverage of generating stub jars without dependencies.
To make that easier, this commit adds an `assertClassesStubbedCorrectly`
method to `StubJarTest` which compares the stub against the relevant parts
of the full jar. (Everything should be identical except that privates,
synthetics, bridges, `<clinit>`, and method bodies are absent, and members of
each type have been sorted.)

Now most tests are simply a matter of a compile call and
an assert, making more tests super easy to add. With this approach, it
no longer makes sense to do the binary comparison of the stub jars, so the
test is now more cleanly parameterized as jar-based, source-based, and
source-based with missing dependencies.

Test Plan: Introduced a few failures, verified that tests correctly fail in response.

Reviewed By: dreiss

fbshipit-source-id: 0f586bb