Add likely-active-network to Buck scuba logging
Summary:
An extra dimension to take a look at build times along. There's a chance that wifi vs ethernet is an interesting social hack for overall build metrics; this allows us to test the assertion at least.

https://pxl.cl/6Kl1

managerswritingdiffs

Test Plan: Build things a few times on devserver and on mac

Reviewed By: marcinkosiba

fbshipit-source-id: e965cef