Use only one headers map that reference all the headers that are going to be used by the target.
Summary:
The size of the build setting HEADER_SEARCH_PATHS in the .xcconfig
when generatin the Xcode project depends on the number of dependencies of the
target.
We'd like to provide an option to keep it short.
This patch introduce an option that will reference all the headers that are
going to be used by the target in a single headers map file.

--config --config apple.merge_header_maps_in_xcode=True

Test Plan:
```
$ buck project --config apple.merge_header_maps_in_xcode=True [myapp]
$ buck project --config apple.merge_header_maps_in_xcode=False [myapp]
```

Reviewed By: yiding

fbshipit-source-id: 74fb189