Ignore OverlappingFileLockException
Summary:
Obtaining the file lock is not synchronized, so multiple threads
can try acquiring it concurrently. Rather than introduce a global
synchronization object and risk unnecessary blocking we can
allow both threads to race for the lock and ignore the error on
one of them.

Test Plan: Couldn't repro issue, this is a speculative fix.

Reviewed By: aiked

fbshipit-source-id: 9e26858