Make the zeroFilesHaveChanged event more precise
Summary:
The `watchmanZeroFileChanges` is firing a bit too much
because it's coming from the private postEvents which doesn't have
full view of all of watchman (just a single cell). In multi-cell
projects, this means we'll see the zero-file-changed notification a
lot more often than is correct.

Test Plan: Add a test, just not today.

Reviewed By: aiked

fbshipit-source-id: 368c449