AndroidPrebuiltAarDescription implements Flavored
Summary:
IntelliJ project file generation works out all of the BuildTargets which needed to be built in order for the project to work, and manually triggers a Build of them. Top-level Builds can only be triggered for flavored things if they implement Flavored.

I haven't implemented the FlavorDomains thing because I don't know what it is and it has a default implementation.

This fixes https://github.com/facebook/buck/issues/1178

Test Plan: None.

Reviewed By: Coneko

fbshipit-source-id: 69ed660